refactor: move decode implementation in its own mod
